backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
openai>=1.0
sentence-transformers
faiss-cpu # Or faiss-gpu if you have CUDA configured
numpy
requests # Add back
langdetect # Add back